Move classes JGraphtTools and FlowNetwork to new sub-package org.mastodon.mamut.treesimilarity.util
