+++ src/main/java/org/mastodon/mamut/treesimilarity/ZhangUnorderedTreeEditDistance.java
@@ -3,6 +3,7 @@ package org.mastodon.mamut.treesimilarity;
 import org.apache.commons.lang3.tuple.Pair;
 import org.mastodon.mamut.treesimilarity.tree.Tree;
 import org.mastodon.mamut.treesimilarity.tree.TreeUtils;
+import org.mastodon.mamut.treesimilarity.util.FlowNetwork;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
