+++ src/test/java/org/mastodon/mamut/treesimilarity/util/JGraphtToolsTest.java
@@ -1,9 +1,8 @@
-package org.mastodon.mamut.treesimilarity;
+package org.mastodon.mamut.treesimilarity.util;
 
 import org.jgrapht.graph.DefaultWeightedEdge;
 import org.jgrapht.graph.SimpleDirectedWeightedGraph;
 import org.junit.Test;
-import org.mastodon.mamut.treesimilarity.JGraphtTools;
 
 import java.util.HashMap;
 import java.util.Map;
